修改了回复异常处理

+++ src/main/java/communtiy/controller/publicController.java
@@ -28,7 +28,7 @@ public class publicController {
 	private QuestionService questionService;
 	
 	@GetMapping("/publish/{id}")
-	public String edit(@PathVariable(name="id")Integer id,Model model) {
+	public String edit(@PathVariable(name="id")Long id,Model model) {
 		QuestionDTO question = questionService.getById(id);
 		model.addAttribute("title", question.getTitle());
 		model.addAttribute("description", question.getDescription());
@@ -47,7 +47,7 @@ public class publicController {
 			@RequestParam("title")String title,
 			@RequestParam("description")String description,
 			@RequestParam("tag")String tag,
-			@RequestParam("id")Integer id,
+			@RequestParam("id")Long id,
 			HttpServletRequest request,
 			Model model) {
